+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## 0.5.62 (Pending):
+* Fix bug that required that the WebPoller factory be setup prior to creating
+  `GwtWebPollerDataLoaderService`.
 
 ## 0.5.61:
 * Add WebPollerDataLoaderService.getWebPollerLogLevel() template method to configure