+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.5.62 (Pending):
+## 0.5.62:
 * Fix bug that required that the WebPoller factory be setup prior to creating
   `GwtWebPollerDataLoaderService`.
 * Introduce `GwtWebPollerDataLoaderService.newRequestBuilder()` template method for